+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/JavascriptExtension.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2024 DiffPlug
+ * Copyright 2016-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,7 +30,6 @@ import org.gradle.api.Project;
 
 import com.diffplug.common.collect.ImmutableList;
 import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.biome.BiomeFlavor;
 import com.diffplug.spotless.npm.EslintConfig;
 import com.diffplug.spotless.npm.EslintFormatterStep;
 import com.diffplug.spotless.npm.NpmPathResolver;
@@ -63,11 +62,9 @@ public class JavascriptExtension extends FormatExtension {
 			extends NpmStepConfig<EslintBaseConfig<T>> {
 		Map<String, String> devDependencies = new LinkedHashMap<>();
 
-		@Nullable
-		Object configFilePath = null;
+		@Nullable Object configFilePath = null;
 
-		@Nullable
-		String configJs = null;
+		@Nullable String configJs = null;
 
 		public EslintBaseConfig(Project project, Consumer<FormatterStep> replaceStep,
 				Map<String, String> devDependencies) {
@@ -169,7 +166,7 @@ public class JavascriptExtension extends FormatExtension {
 		 * @param version Biome version to use.
 		 */
 		public BiomeJs(String version) {
-			super(getProject(), JavascriptExtension.this::replaceStep, BiomeFlavor.BIOME, version);
+			super(getProject(), JavascriptExtension.this::replaceStep, version);
 		}
 
 		@Override